Revert "Allow multiple hooks objects, create types file"
