+++ src/hooks/logEvent.ts
@@ -1,4 +1,4 @@
-import { HookCreator } from '../types/hooks'
+import { HookCreator } from '../index'
 
 export const logEvent: HookCreator = () => async state => {
     console.log(`received event: ${state.event}`)
