+++ src/index.ts
@@ -1,9 +1,5 @@
 import { Context } from 'aws-lambda'
 
-import { Event, HooksObject, UseHooks, UseHooksState } from './types/hooks'
-import { combineHooks } from './utils'
-import { defaultHook } from './utils/defaultHook'
-
 export {
     handleScheduledEvent,
     handleUnexpectedError,
@@ -12,28 +8,62 @@ export {
     validateEventBody,
 } from './hooks'
 
-export * from './types/hooks'
+interface Hooks {
+    before?: HookHandler[]
+    after?: HookHandler[]
+    onError?: HookHandler[]
+}
+
+type Response = any
+type Event = any
+
+interface State {
+    event: Event
+    context: Context
+    exit: boolean
+    response?: Response
+    error?: Error
+}
 
 /**
- * Using the provided hooks create an `withHooks` higher order function
- * @param hooks a variadic array of config objects containing the hooks to apply to your lambda. Each config object can contain `before`, `after` and `onError` arrays
+ * @param config optional configuration object for this hook
+ * @returns HookHandler
+ */
+export type HookCreator<Config = {}> = (config?: Config) => HookHandler
+/**
+ * @param state a state object that might be manipulated by this function
+ * @param state.event event passed in from AWS
+ * @param state.context context passed in from AWS
+ * @param state.exit defaults to false, if set to true program will exit early after ivocation of this hook
+ * @param state.response returned when state.exit is set to true
+ * @param state.error exists only if there's an unhandled exception thrown inside a hook or the lambda
+ * @returns Promise<state>
+ */
+type HookHandler = (state: State) => Promise<State>
+
+type UseHooks = (hooks: Hooks) => WithHooks
+type WithHooks = (lambda: any) => (event: any, context: Context) => Promise<any>
+/**
+ * Using the provided hooks create an withHooks higher order function
+ * @param hooks a config object of the hooks to apply to your lambda
+ * @param hooks.before an array of hooks to run before the provided lambda
+ * @param hooks.after an array of hooks to run after the provided lambda
+ * @param hooks.onError an array of hooks to run only if there's an error during the execution
  * @returns WithHooks() function that wraps around your lambda
  */
-export const useHooks: UseHooks = (...hooksArr) => {
-    if (!hooksArr) {
-        hooksArr = [defaultHook]
-    }
-
-    const hooks: HooksObject = combineHooks(hooksArr)
+export const useHooks: UseHooks = (hooks: Hooks): WithHooks => {
+    if (!hooks.before) hooks.before = []
+    if (!hooks.after) hooks.after = []
+    if (!hooks.onError) hooks.onError = []
 
     /**
      * Higher order function that takes a lambda function
      * as input and applies the hooks provided to useHooks()
      * @param lambda lambda function
-     * @returns supercharged lambda 🚀
+     * @returns supercharged lambda  🚀
      */
     const withHooks = (lambda: any) => async (event: Event, context: Context) => {
-        let state: UseHooksState = { event, context, exit: false }
+        let state: State = { event, context, exit: false }
 
         try {
             for (const hook of hooks.before!) {
