--- src/utils/defaultHook.ts
@@ -1,7 +1,0 @@
-import { HooksObject } from '../types/hooks'
-
-export const defaultHook: HooksObject = {
-    before: [],
-    after: [],
-    onError: [],
-}
